Sprint 15

+++ Tracker/TrackerViewController/Models/TrackerRecord.swift
@@ -1,7 +1,6 @@
-
 import Foundation
 
 struct TrackerRecord {
     let id: UUID
-    let date: UInt
+    let date: UInt64
 }
